+++ src/Fp/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - WriteContext extensions in ReadWriteContextByteExtensions
 - `WriteContext<byte>.GetAdvance(int)` to claim buffer
+- Clearing of stored singletons via `One<T>.Clear` / and `One.ClearAll`
+- `StringData` type for data containing string
+- `Data.SupportsFormat` and `Data.SupportedFormats` for pre-checking output support
 
 ### Changed
 - Fix xor arm intrinsic